requirements.txt
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
enum34==1.1.6
fire==0.1.3
futures==3.1.1
google-api-core==1.2.1
google-assistant-grpc==0.2.0
google-assistant-library==1.0.0
google-assistant-sdk==0.5.0
google-auth==1.5.0
google-auth-oauthlib==0.2.0
google-cloud-vision==0.32.0
googleapis-common-protos==1.5.3
grpcio==1.12.1
idna==2.7
imutils==0.4.6
numpy==1.14.5
oauthlib==2.1.0
pathlib2==2.3.2
picamera==1.13
Pillow==5.1.0
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pytz==2018.4
pyzbar==0.1.7
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
sounddevice==0.3.11
tenacity==4.12.0
urllib3==1.23